Move generalizable code out of bluenop.py
